src/tvm-job-clean/requirements.txt
google-cloud-compute==1.3.0